+++ app/services/bloom_trade_client/exchange_rates/convert.rb
@@ -6,13 +6,14 @@ module BloomTradeClient
         base_currency:,
         counter_currency:,
         reserve_currency: BloomTradeClient.configuration.reserve_currency,
-        type: "mid"
+        type: "mid",
+        jwt:
       )
-        rate = direct_rate(type, base_currency, counter_currency)
+        rate = direct_rate(type, base_currency, counter_currency, jwt)
         return rate if rate
 
-        origin_rate = direct_rate(type, base_currency, reserve_currency)
-        destination_rate = direct_rate(type, reserve_currency, counter_currency)
+        origin_rate = direct_rate(type, base_currency, reserve_currency, jwt)
+        destination_rate = direct_rate(type, reserve_currency, counter_currency, jwt)
         reverse_rate = origin_rate.rate * destination_rate.rate if origin_rate && destination_rate
         return ConversionResult.new(rate: reverse_rate) if reverse_rate
 
@@ -24,14 +25,16 @@ module BloomTradeClient
 
       private
 
-      def self.direct_rate(type, base_currency, counter_currency)
+      def self.direct_rate(type, base_currency, counter_currency, jwt)
         return nil unless %w(buy sell mid).include? type
-
         return ConversionResult.new(rate: 1.0) if base_currency == counter_currency
+
+        jwt_hash = jwt ? Digest::SHA256.base64digest(jwt) : nil
 
         exchange_rate = ExchangeRate.find_by(
           base_currency: base_currency,
           counter_currency: counter_currency,
+          jwt_hash: jwt_hash
         )
 
         if exchange_rate
@@ -44,6 +47,7 @@ module BloomTradeClient
         reversed_rate = ExchangeRate.find_by(
           base_currency: counter_currency,
           counter_currency: base_currency,
+          jwt_hash: jwt_hash
         )
 
         if reversed_rate