+++ lib/bloom_trade_client.rb
@@ -17,7 +17,8 @@ module BloomTradeClient
 
   include APIClientBase::Base.module
 
-  EXCHANGE_RATES_CHANNEL = '/exchange_rates'.freeze
+  RATES_CHANNEL     = '/exchange_rates'.freeze
+  ORG_RATES_CHANNEL = '/org_exchange_rates'.freeze
 
   with_configuration do
     has(:host, {
@@ -25,13 +26,15 @@ module BloomTradeClient
       default: 'https://staging.trade.bloom.solutions',
     })
     has :reserve_currency, classes: String, default: 'PHP'
+    has :jwt_callback, classes: Object
   end
 
-  def self.convert(base_currency:, counter_currency:, type:)
+  def self.convert(base_currency:, counter_currency:, type:, jwt: nil)
     BloomTradeClient::ExchangeRates::Convert.(
       base_currency: base_currency,
       counter_currency: counter_currency,
-      type: type
+      type: type,
+      jwt: jwt
     )
   end
 end