+++ spec/services/bloom_trade_client/exchange_rates/sync_spec.rb
@@ -33,18 +33,34 @@ module BloomTradeClient
           "expires_at" => Time.current.to_i,
         }
       end
+      let(:data_4) do
+        {
+          "base_currency" => "USD",
+          "counter_currency" => "PHP",
+          "buy" => "70",
+          "mid" => "80",
+          "sell" => "90",
+          "expires_at" => Time.current.to_i,
+        }
+      end
+      let(:token) { "jwt-token" }
+      let(:headers) do
+        { "Authorization" => "Bearer #{token}" }
+      end
 
       it "receives a payload and saves or updates the Exchange Rate" do
         [data_1, data_2, data_3].each do |data|
-          described_class.(data, nil)
+          described_class.(data, nil, nil)
         end
+        described_class.(data_4, nil, headers)
 
         expect_created_rates = BloomTradeClient::ExchangeRate.all
-        expect(expect_created_rates.count).to eq 3
+        expect(expect_created_rates.count).to eq 4
 
         btcusd = BloomTradeClient::ExchangeRate.find_by(
           base_currency: "BTC",
           counter_currency: "USD",
+          jwt_hash: nil
         )
 
         expect(btcusd).to be_present
@@ -56,6 +72,7 @@ module BloomTradeClient
         audusd = BloomTradeClient::ExchangeRate.find_by(
           base_currency: "AUD",
           counter_currency: "USD",
+          jwt_hash: nil
         )
 
         expect(audusd).to be_present
@@ -67,6 +84,7 @@ module BloomTradeClient
         usdphp = BloomTradeClient::ExchangeRate.find_by(
           base_currency: "USD",
           counter_currency: "PHP",
+          jwt_hash: nil
         )
 
         expect(usdphp).to be_present
@@ -76,7 +94,7 @@ module BloomTradeClient
         expect(usdphp.expires_at).to eq data_3['expires_at']
       end
 
-      context "the exchange rates currently exist" do
+      context "the global exchange rates currently exist" do
         before do
           create(:bloom_trade_client_exchange_rate, {
             base_currency: "BTC",
@@ -84,20 +102,23 @@ module BloomTradeClient
             buy: 7800,
             mid: 7900,
             sell: 8000,
+            jwt_hash: nil,
           })
         end
 
         it "updates the old rate" do
           [data_1, data_2, data_3].each do |data|
-            described_class.(data, nil)
+            described_class.(data, nil, nil)
           end
+          described_class.(data_4, nil, headers)
 
           expect_created_rates = BloomTradeClient::ExchangeRate.all
-          expect(expect_created_rates.count).to eq 3
+          expect(expect_created_rates.count).to eq 4
 
           btcusd = BloomTradeClient::ExchangeRate.find_by(
             base_currency: "BTC",
             counter_currency: "USD",
+            jwt_hash: nil
           )
 
           expect(btcusd).to be_present
@@ -105,6 +126,44 @@ module BloomTradeClient
           expect(btcusd.mid).to eq 8000
           expect(btcusd.sell).to eq 8100
           expect(btcusd.expires_at).to eq data_1['expires_at']
+        end
+      end
+
+      context "the jwt-scoped exchange rate currently exist" do
+        let(:jwt_hash) { Digest::SHA256.base64digest(token) }
+
+        before do
+          create(:bloom_trade_client_exchange_rate, {
+            base_currency: "USD",
+            counter_currency: "PHP",
+            buy: "100",
+            mid: "110",
+            sell: "120",
+            expires_at: Time.current.to_i,
+            jwt_hash: jwt_hash,
+          })
+        end
+
+        it "updates the old rate" do
+          [data_1, data_2, data_3].each do |data|
+            described_class.(data, nil, nil)
+          end
+          described_class.(data_4, nil, headers)
+
+          expect_created_rates = BloomTradeClient::ExchangeRate.all
+          expect(expect_created_rates.count).to eq 4
+
+          btcusd = BloomTradeClient::ExchangeRate.find_by(
+            base_currency: "USD",
+            counter_currency: "PHP",
+            jwt_hash: jwt_hash
+          )
+
+          expect(btcusd).to be_present
+          expect(btcusd.buy).to eq 70
+          expect(btcusd.mid).to eq 80
+          expect(btcusd.sell).to eq 90
+          expect(btcusd.expires_at).to eq data_4["expires_at"]
         end
       end