+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.18.0] - 2018-02-12
+### Added
+- New `jwt_hash` column for `ExchangeRate`
+- New `jwt_callback` config which accepts an Object that responds to 
+`.call`.
+- New `SyncJob` that fetches rates globally and per JWT returned 
+- `BloomTradeClient::Convert` accepts an optional `jwt`.
+by `jwt_callback`
+
 ## [0.17.0] - 2018-12-05
 ### Added
 - Add `.errors` to `BaseResponse`